add seperate fav and all movie components and containers
