+++ app/index.js
@@ -2,7 +2,9 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import {createStore} from 'redux';
 import App from './containers/App-container';
-import MovieIndex from './containers/MovieIndex-container';
+import AllMovies from './containers/AllMovies-container';
+import FavMovies from './containers/FavMovies-container';
+// import MovieIndex from './containers/MovieIndex-container';
 import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import allReducers from './reducers/index';
 import Login from './components/Login';
@@ -15,9 +17,9 @@ const router = (
   <Provider store = {store}>
     <Router history={browserHistory} >
       <Route path='/' component={App}>
-        <IndexRoute component={MovieIndex} />
+        <IndexRoute component={AllMovies} />
         <Route path='/login' component={Login} />
-        <Route path='/favorites' component={MovieIndex} />
+        <Route path='/favorites' component={FavMovies} />
       </Route>
     </Router>
   </Provider>
