+++ app/reducers/reducer-movie.js
@@ -1,7 +1,9 @@
 const movieReducer = (state = [], action) => {
   switch(action.type){
     case 'ADD_MOVIES':
-      return action.movies
+      return Object.assign({}, state, {movies: action.movies})
+    case 'FINAL_FAVS':
+      return Object.assign({}, state, {final: action})
   }
   return state;
 }
